riscv: bakewell: vdso: Build with --no-cap-table-relocs
Build the startup pi code and the vdso with --no-cap-table-relocs.
Additionally, use capability register in asm register statements
in the vdso if those reference pointers.

+++ arch/riscv/include/asm/vdso.h
@@ -20,7 +20,7 @@
 #include <generated/vdso-offsets.h>
 
 #define VDSO_SYMBOL(base, name)							\
-	(void __user *)((unsigned long)(base) + __vdso_##name##_offset)
+	(void __user *)((uintptr_t)(base) + (unsigned long)__vdso_##name##_offset)
 
 #ifdef CONFIG_COMPAT
 #include <generated/compat_vdso-offsets.h>
